[not verified] move init state to another class
